+++ truffle/CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Removed deprecated ResultVerifier.getDefaultResultVerfier.
 * Deprecated `com.oracle.truffle.api.frame.FrameDescriptor.shallowCopy` and `com.oracle.truffle.api.frame.FrameSlot.getFrameDescriptor` 
 * Added [DebugValue#set](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/debug/DebugValue.html#set-java.lang.Object-) to set primitive values to a debug value.
+* Added support for [logging](http://www.graalvm.org/truffle/javadoc/com/oracle/truffle/api/Truffle.html#getLogger-java.lang.String-java.lang.Class-) in Truffle languages and instruments.
 
 ## Version 1.0.0 RC2